V2

+++ extensions/network/HttpResponse.h
@@ -110,7 +110,7 @@ public:
     /** Get the http response errorCode
      *  I know that you want to see http 200 :)
      */
-    inline int getResponseCode()
+    inline long getResponseCode()
     {
         return _responseCode;
     }
@@ -153,7 +153,7 @@ public:
     
     /** Set the http response errorCode
      */
-    inline void setResponseCode(int value)
+    inline void setResponseCode(long value)
     {
         _responseCode = value;
     }
@@ -175,7 +175,7 @@ protected:
     bool                _succeed;       /// to indecate if the http reqeust is successful simply
     std::vector<char>   _responseData;  /// the returned raw data. You can also dump it as a string
     std::vector<char>   _responseHeader;  /// the returned raw header data. You can also dump it as a string
-    int                 _responseCode;    /// the status code returned from libcurl, e.g. 200, 404
+    long                 _responseCode;    /// the status code returned from libcurl, e.g. 200, 404
     std::string         _errorBuffer;   /// if _responseCode != 200, please read _errorBuffer to find the reason 
     
 };
